# gdtracker-go-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GDTracker Go API
  version: 0.5.0
  description: |
    Go backend for GDTracker. Session auth and CSRF mirror the Spring API contract (see `docs/AUTH.md`).
    Routes mirror `gdtracker-api` for games, categories, tags, tasks, features, game configuration, archives, game exceptions,
    ingest plane (players, integration ping), game events (definitions + ingest), traces (heatmap), and feedback.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/games/{gameId}/game-feedback-meter-definitions:
    get:
      tags: [GameFeedback]
      operationId: listGameFeedbackMeterDefinitions
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GameFeedbackMeterDefinition" }
        "401": { description: Not authenticated }
        "404": { description: Game not found or not owned }
    post:
      tags: [GameFeedback]
      operationId: createGameFeedbackMeterDefinition
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GameFeedbackMeterDefinitionUpsertRequest" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameFeedbackMeterDefinition" }
        "400": { description: Validation error }
        "401": { description: Not authenticated }
        "404": { description: Game not found or not owned }
        "409": { description: Duplicate field key }
  /api/games/{gameId}/game-feedback-meter-definitions/{id}:
    put:
      tags: [GameFeedback]
      operationId: updateGameFeedbackMeterDefinition
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/FeedbackMeterId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GameFeedbackMeterDefinitionUpsertRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameFeedbackMeterDefinition" }
        "400": { description: Validation error }
        "401": { description: Not authenticated }
        "404": { description: Not found }
        "409": { description: Duplicate field key }
    delete:
      tags: [GameFeedback]
      operationId: deleteGameFeedbackMeterDefinition
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/FeedbackMeterId"
      responses:
        "204": { description: Deleted }
        "401": { description: Not authenticated }
        "404": { description: Not found }
  /api/games/{gameId}/game-feedback:
    get:
      tags: [GameFeedback]
      operationId: listGameFeedback
      parameters:
        - $ref: "#/components/parameters/GameId"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GameFeedbackSummary" }
        "401": { description: Not authenticated }
        "404": { description: Game not found or not owned }
    post:
      tags: [GameFeedback]
      operationId: ingestGameFeedback
      security: []
      parameters:
        - $ref: "#/components/parameters/GameId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GameFeedbackIngestRequest" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameFeedback" }
        "400": { description: Validation error }
        "401": { description: Invalid token or player }
        "404": { description: Game not found }
  /api/games/{gameId}/game-feedback/{feedbackId}:
    get:
      tags: [GameFeedback]
      operationId: getGameFeedbackDetail
      parameters:
        - $ref: "#/components/parameters/GameId"
        - $ref: "#/components/parameters/FeedbackId"
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GameFeedbackDetail" }
        "401": { description: Not authenticated }
        "404": { description: Not found }
components:
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema: { type: string }
...
    FeedbackMeterId:
      name: id
      in: path
      required: true
      schema: { type: string }
    FeedbackId:
      name: feedbackId
      in: path
      required: true
      schema: { type: string }
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ingest-token
  schemas:
...
    GameFeedbackMeterDefinition:
      type: object
      properties:
        id: { type: string }
        fieldKey: { type: string }
        question: { type: string }
        sortOrder: { type: integer }
    GameFeedbackMeterDefinitionUpsertRequest:
      type: object
      required: [fieldKey, question]
      properties:
        fieldKey: { type: string }
        question: { type: string }
        sortOrder: { type: integer }
    GameFeedbackSummary:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        playerId: { type: string }
        createdAt: { type: string, format: date-time }
    GameFeedbackDetail:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        playerId: { type: string }
        createdAt: { type: string, format: date-time }
        meters:
          type: array
          items: { $ref: "#/components/schemas/GameFeedbackMeterValue" }
    GameFeedbackMeterValue:
      type: object
      properties:
        fieldKey: { type: string }
        question: { type: string }
        value: { type: integer }
    GameFeedbackIngestRequest:
      type: object
      required: [title, description]
      properties:
        title: { type: string }
        description: { type: string }
        meters:
          type: object
          additionalProperties: { type: integer }
    GameFeedback:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        playerId: { type: string }
        createdAt: { type: string, format: date-time }
        meters:
          type: object
          additionalProperties: { type: integer }